rename: 관련 컴포넌트 section 폴더로 이동
