+++ src/app/main/_components/content/MainContent.tsx
@@ -1,0 +1,5 @@
+const MainContent = () => {
+  return <main className="flex flex-col gap-4"></main>;
+};
+
+export default MainContent;
